requirements.txt
flask==2.3.0
python-docx==1.1.0
pdfplumber==0.9.0
opencv-python>=4.8.0.76
numpy>=1.24.0
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.1
requests==2.31.0
python-dotenv==1.0.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.2
openai-whisper
scikit-learn>=1.3.0
Pillow>=10.0.0
